refactor(test): converge tool action fixture into shared test_fixtures
The same tool Action builder existed in picker.rs, tui/content/tool.rs,
and tui/content/block.rs. Move it to src/test_fixtures::tool_action_part;
picker uses it directly (11 call sites unchanged) and the TUI files
become one-line delegates over it.
